0.1.0 target 集成

+++ SensorsAnalyticsSDK/AutoTrackUtils.m
@@ -60,7 +60,7 @@
         if (sameTypeViews.count > 1) {
             NSString * className = nil;
             NSUInteger index = [sameTypeViews indexOfObject:view];
-            className = [NSString stringWithFormat:@"%@[%lu]", NSStringFromClass([view class]), index];
+            className = [NSString stringWithFormat:@"%@[%lu]", NSStringFromClass([view class]), (unsigned long)index];
             [viewPathArray addObject:className];
         } else {
             [viewPathArray addObject:NSStringFromClass([view class])];
@@ -110,7 +110,7 @@
             if (sameTypeViews.count > 1) {
                 NSString * className = nil;
                 NSUInteger index = [sameTypeViews indexOfObject:responder];
-                className = [NSString stringWithFormat:@"%@[%lu]", NSStringFromClass([responder class]), index];
+                className = [NSString stringWithFormat:@"%@[%lu]", NSStringFromClass([responder class]), (unsigned long)index];
                 [viewPathArray addObject:className];
             } else {
                 [viewPathArray addObject:NSStringFromClass([responder class])];
@@ -137,7 +137,7 @@
                     if (items.count > 1) {
                         NSString * className = nil;
                         NSUInteger index = [items indexOfObject:responder];
-                        className = [NSString stringWithFormat:@"%@[%lu]", NSStringFromClass([responder class]), index];
+                        className = [NSString stringWithFormat:@"%@[%lu]", NSStringFromClass([responder class]), (unsigned long)index];
                         [viewPathArray addObject:className];
                     } else {
                         [viewPathArray addObject:NSStringFromClass([responder class])];
@@ -345,7 +345,7 @@
                 [viewPathArray addObject:NSStringFromClass([responder class])];
             } else {
                 NSInteger index = [viewsArray indexOfObject:collectionView];
-                [viewPathArray addObject:[NSString stringWithFormat:@"%@[%ld]", NSStringFromClass([responder class]), index]];
+                [viewPathArray addObject:[NSString stringWithFormat:@"%@[%ld]", NSStringFromClass([responder class]), (long)index]];
             }
             
             responder = [responder nextResponder];
@@ -506,7 +506,7 @@
                 [viewPathArray addObject:NSStringFromClass([responder class])];
             } else {
                 NSInteger index = [viewsArray indexOfObject:tableView];
-                [viewPathArray addObject:[NSString stringWithFormat:@"%@[%ld]", NSStringFromClass([responder class]), index]];
+                [viewPathArray addObject:[NSString stringWithFormat:@"%@[%ld]", NSStringFromClass([responder class]), (long)index]];
             }
             responder = [responder nextResponder];
             [self sa_find_responder:responder withViewPathArray:viewPathArray];
